% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_gazetadoacre.R
\name{baixar_gazetadoacre}
\alias{baixar_gazetadoacre}
\title{Baixar Gazeta do Acre}
\usage{
baixar_gazetadoacre(aba = "policia", periodo = "", paginas = 10,
  diretorio = ".")
}
\arguments{
\item{aba}{nome da aba, e.x, "policia"}

\item{periodo}{colocar "agora" para notícias que estão no ar agora,
ignorar se quiser informar o número de páginas.}

\item{paginas}{Se período ignorado, informar o número de páginas
a serem baixadas}

\item{diretorio}{Diretório onde salvar os htmls}
}
\value{
htmls baixados
}
\description{
Baixar Gazeta do Acre
}
\examples{
\dontrun{
baixar_gazetadoacre(aba = "policia", periodo = "agora")
}
}
